refactor(catalog): move CatalogManager to crate behind filesystem-resolver port (Slice 2)
Second increment of the catalog extraction. Moves CatalogManager +
TableOpLockRegistry (the central runtime) from root src/catalog/mod.rs into the
proximadb-catalog crate (manager.rs), dissolving the catalog->storage up-edge
behind a new CatalogFilesystemResolver port (the observatory-Slice-3 pattern).

The port: CatalogManager held the concrete FilesystemFactory (root) for
object-store catalog URLs (s3://, gs://, az://). Now it holds
Option<Arc<dyn CatalogFilesystemResolver>>; the object-store branch calls the
port. The root half (LazyFilesystemResolver) wraps FilesystemFactory in a
tokio OnceCell -- the factory is created lazily on first object-store catalog
URL (never for local-only setups), preserving the original behavior. Injected
at the composition root (shared_services) via set_filesystem_resolver;
CatalogManager::new() stays argless so no caller/test signature ripples.

Recon (code-graph Martin metrics, now fixed): src/catalog/mod.rs was ca=54,
ce=20, instability=0.27 -- only ONE real up-edge (FilesystemFactory in
create_native_catalog); the cluster::partition_lease ref was doc-only.

Root src/catalog/mod.rs -> re-export shim (CatalogManager/TableOpLockRegistry/
CatalogFilesystemResolver from the crate + the new filesystem_resolver module).
All crate::catalog::CatalogManager paths unchanged. The ~36 CatalogManager
tests stay in root (they test the crate type via re-export); only the `Arc`
import moved into the test mod.

Verification: crate + root clippy --lib --bins -D warnings clean; fmt clean;
layering clean (storage up-edge ratchet unchanged at 197); 230 catalog lib
tests pass.
